Edited Token unit and integration tests to get correct creat tests in right places and improve scenario coverage.
